archive/lab5/requirements_updated.txt
tornado>=6.0
numpy>=1.21.0
opencv-python>=4.5.0
onnxruntime>=1.8.0
Shapely>=1.8.0
pyclipper>=1.3.0
Pillow>=9.0.0





